extract DonationControls from index.js
